+++ frontend/User/package.json
@@ -12,6 +12,8 @@
 		"format": "prettier --plugin-search-dir . --write ."
 	},
 	"devDependencies": {
+		"@felte/reporter-tippy": "^1.1.4",
+		"@felte/validator-yup": "^1.0.10",
 		"@sveltejs/adapter-node": "next",
 		"@sveltejs/kit": "next",
 		"@tailwindcss/nesting": "^0.0.0-insiders.565cd3e",
@@ -24,24 +26,22 @@
 		"eslint": "^8.16.0",
 		"eslint-config-prettier": "^8.3.0",
 		"eslint-plugin-svelte3": "^4.0.0",
+		"felte": "^1.2.6",
 		"postcss": "^8.4.14",
 		"postcss-load-config": "^4.0.1",
 		"postcss-preset-env": "^7.8.2",
 		"prettier": "^2.6.2",
 		"prettier-plugin-svelte": "^2.7.0",
+		"sass": "^1.55.0",
 		"svelte": "^3.44.0",
 		"svelte-check": "^2.7.1",
 		"svelte-preprocess": "^4.10.7",
 		"tailwindcss": "^3.1.5",
+		"tippy.js": "^6.3.7",
 		"tslib": "^2.3.1",
 		"typescript": "^4.7.4",
-		"vite": "^3.1.0"
-	},
-	"type": "module",
-	"dependencies": {
-		"@felte/validator-yup": "^1.0.10",
-		"felte": "^1.2.6",
-		"sass": "^1.55.0",
+		"vite": "^3.1.0",
 		"yup": "^0.32.11"
-	}
+	},
+	"type": "module"
 }
